No lee valores y sale error cuando compilo en el execute

+++ src/entities/signal_pwm.h
@@ -5,16 +5,16 @@
 #include <entities/value.h>
 
 
-class SignalPWM : public ValueABC<uint8_t> {
+class SignalPWM : public ValueABC<int> {
     public:
         SignalPWM():ValueABC(){};
-        SignalPWM(uint8_t value, uint8_t source):ValueABC(value, source){};
-        uint8_t getValue();
+        SignalPWM(int value, uint8_t source):ValueABC(value, source){};
+        int getValue();
     private:
         uint8_t value;
 };
 
-uint8_t SignalPWM::getValue() {
+int SignalPWM::getValue() {
     return this->value;
 }
